# perf: implement new radix 4 NR NTT