Measured locally on M1: 4-5% throughput improvement over radix 2.
Criterion benchmark added.

## Type of change

- [x] Optimization

## Checklist
- [x] This change is an Optimization
  - [x] Benchmarks added/run

## Preliminary Results on Development M1

```
# 2^24 elements
Ordered FFT/Sequential from NR radix2
                        time:   [5.1193 s 5.1273 s 5.1355 s]
                        thrpt:  [3.2669 Melem/s 3.2722 Melem/s 3.2773 Melem/s]
Ordered FFT/Sequential from NR radix4
                        time:   [4.8796 s 4.8917 s 4.9056 s]
                        thrpt:  [3.4200 Melem/s 3.4297 Melem/s 3.4382 Melem/s]
```

## Results on `lambdaworks-benchmarks-x86-64-dedicated`
```
# 2^20 elements
Ordered FFT/Sequential from NR radix2
                        time:   [334.31 ms 334.76 ms 335.27 ms]
                        thrpt:  [3.1276 Melem/s 3.1324 Melem/s 3.1365 Melem/s]
Ordered FFT/Sequential from NR radix4
                        time:   [307.61 ms 307.75 ms 307.88 ms]
                        thrpt:  [3.4058 Melem/s 3.4073 Melem/s 3.4088 Melem/s]
# 2^22 elements
Ordered FFT/Sequential from NR radix2
                        time:   [1.4725 s 1.4734 s 1.4744 s]
                        thrpt:  [2.8447 Melem/s 2.8466 Melem/s 2.8484 Melem/s]
Ordered FFT/Sequential from NR radix4
                        time:   [1.3526 s 1.3534 s 1.3544 s]
                        thrpt:  [3.0968 Melem/s 3.0991 Melem/s 3.1010 Melem/s]
# 2^24 elements
Ordered FFT/Sequential from NR radix2
                        time:   [6.4369 s 6.4393 s 6.4419 s]
                        thrpt:  [2.6044 Melem/s 2.6054 Melem/s 2.6064 Melem/s]
Ordered FFT/Sequential from NR radix4
                        time:   [5.9197 s 5.9247 s 5.9301 s]
                        thrpt:  [2.8292 Melem/s 2.8317 Melem/s 2.8341 Melem/s]
```